backend/requirements.txt
# FastAPI 和相关依赖
fastapi==0.95.0
uvicorn==0.21.0

# Pydantic v2 - 使用当前版本
pydantic==1.10.7

# 数据库相关
sqlalchemy==1.4.41
psycopg2-binary==2.9.6

# HTTP 客户端
httpx==0.24.0

# 环境变量
python-dotenv==0.21.1
python-multipart==0.0.6
alembic==1.7.7
email-validator==2.0.0
python-jose==3.3.0 